Dialog, Label, Message: use useI18nWithOverride composable
This i18n fix ensures that when the customizable prop is set, the
Dialog/Message button's aria label or Label's optional flag is
set to the prop's value. Otherwise, when the prop is not set,
the translated message or fallback value is set.

The customizable props utilize the `useI18nWithOverride` composable.

Tests were updated to reflect these changes.

Bug: T370555
Change-Id: Ieb6f8429f9a97e3b48d27ec1ae88856d29917799