FloatingUI: Use default fallbackStrategy
When FloatingUI attempts to find a placement for an element
and cannot, it uses a fallback strategy to place the menu. We've
been using `initialPlacement`, which falls back to the initial
placement specified for the element. For MenuButton, we use
`bottom-start`, for all other Menu components, we use `bottom`,
and for Tooltip it varies.

In specific layouts, this cases issues for MenuButton, and the
menu can end up opening down/to the start of the triggering
element even when that is far from ideal. It also doesn't
seem like the right choice for other components that use
FloatingUI either. So, this patch removes the fallbackStrategy
in order to use the default strategy, `bestFit`.

See https://floating-ui.com/docs/flip#fallbackstrategy

Bug: T369934
Change-Id: Ia4b6df400d7ecf8f7ccc2c3005e958e34125c59f