docs: update releasing docs
Refine the MediaWiki section to include pulling down the latest
version of master, installing dependencies, and creating a new
branch to work on. And update the language in the PatchDemo
section.

When you don’t update the dependencies, expect weird behavior.
For example: When you run the script to generate a new integrity
hash, no integrity hash is generated.

Change-Id: I995e80ef730cff2aff09f4f9195e7cc0b123b347