Table: Server-side pagination
Introduces the following new props and events to support
pagination of remote data:

Props
* pending (boolean)
* serverPagination (boolean)
* totalRows (number, optional)

Events
* load-more
* last

i18n Messages
* cdx-table-pagination-status-message-indeterminate-short
* cdx-table-pagination-status-message-indeterminate-long
* cdx-table-pagination-status-message-indeterminate-final

Server-side pagination is activated by setting the
"serverPagination" prop to "true", and optionally providing
a "totalRows" prop (for a better UX). At this point the table
will set the default page size to the number of rows that it
has been initially provided with, and any user interaction
with the pagination controls will fire a "load-more" event.

Typical usage will probably look like this:
* the data passed down to the Table is stored as a ref in
  the parent component, and passed down as a prop
* parent component is wired to listen for "load-more" events
  and initiates an API request when the event is fired. The
  payload of this event can be used to determine which results
  to retrieve and how many to ask for
* During this time, the "pending" prop can be set to true so
  that the table activates its pending state and makes it clear
  to the user that new data is incoming
* Once new data is received from the API, the value of the
  original ref is replaced (updating the table's contents) and
  the pending prop is set back to false.

Bug: T370445
Change-Id: If7e68d917f5e73eb122bcdfc31b6ef91f155f5e0