Table: Fix sort icon spacing
The sort icon previously had a padding-left (or right, depending
on the column alignment) to separate it from the column heading
text. When box-sizing is set to border-box, this caused the icon
itself to shrink, since the padding was taken into account when
calculating the size of the icon.

This patch changes the padding to margin to avoid this issue.

Bug: T370959
Change-Id: I9d229ec8e1c5995403f02fc34326e52ec9b06a9b