useI18n: Use WatchSource instead of MaybeRef for parameters
A WatchSource can be a ref or a function that returns a value. This
ensures that the parameters are always reactive, unlike MaybeRef (which
can accept non-reactive plain values).

This makes the code for the cdx-table-select-row-label message a bit
annoying (it's one of the rare cases where passing a plain value is
actually fine), but I think that's worth it to prevent ourselves from
making reactivity mistakes.

Change-Id: I4f5a04631f40881757b9502f1269242b9864c54e