docs: Add support for @default tag for props, and use in Table
For props with complicated default values, vue-docgen-cli doesn't
produce very helpful output: it's just the function that generates the
default value. Add support for an optional `@default` tag that we can
use to override this when vue-docgen-cli infers something unhelpful.

Unfortunately, we also need to use this for almost all array and object
parameters, to change the inferred default from e.g. `() => []` to `[]`.
vue-docgen-cli does try to infer that automatically, but that inference
only works if you use a return statement in the arrow function (e.g.
`() => { return []; }`) which we generally don't, and if the type of the
prop is 'array' or 'object' which it almost never is because we use
TypeScript types.

In a follow-up change we should go through the other components and
add @default tags to the props that need it.

Change-Id: I6df71bb2d72792c8bab67f71864945b9c5e8ed44