i18n: Build a JSON file for use in MW from Codex message keys
This patch adds a new entry point, "src/messages.ts", which is
used in a dedicated library build script. This file reads the
value of the I18nMessageKeys constant and writes the data to
a dist/messageKeys.json file as part of the Codex build process.
It does not otherwise participate in the Codex codebase at all
and is thus ignored by Vite, tsconfig, jest, etc.

The tsx library has also been added as a devDependency, because
that seems to be the most straightforward way to execute a TS file
from the Node.js command line without having to first compile it
or set up a larger scaffold.

Bug: T371330
Change-Id: I203eb9663bc2e1ba80b523538ba97ca92f71fcb7