docs: Use Codex colors for links within content
Y'all...we weren't actually using Codex colors for links on the
docs site. VitePress doesn't use the CSS custom properties we
expected it to for link colors, and we weren't providing an
override for the custom properties it does use. This patch fixes
that, and add a Codex color for active links, which VitePress
doesn't set.

Change-Id: I1f7f0807f186daa12936cf96b40b31f84c74a8a9