Link: Don't accidentally style thumbnail placeholder icons
The Link mixin includes styles for external icons by targeting
`.cdx-icon:last-child`. This assumes that anything with this
relatively generic selector is an external link icon, which is
not the case in Vector 2022. This patch fixes the immediate
problem, which is that placeholder icons in TypeaheadSearch
results are styled inappropriately (off-center and too-small).

Eventually, we should completely rethink the way we target
external icons and change this code to protect against other
unwanted consequences. This would be a breaking change, so
we should consider it for Codex 2.0.

Bug: T372420
Change-Id: I7eb9dcff42b2a19601a8229b7a3ddb2472a2c842