Tabs: make 2-way binding of active tab optional
This change updates the Tabs component to use the newly-introduced
`useOptionalModelWrapper` composable to make 2-way binding optional.

Standard usage of the Tabs component can now work without `v-model:active`
at all; the first tab is set as active on render by default if this is
not provided. Docs and demos have been updated to prefer the simpler
usage.

2-way binding is still available and will make the most sense
when the active tab is meant to be bound to a URL parameter or fragment.

Change-Id: Icf24a1531b45449fe6ee7b0a33328b2092f50849