Message: Set icon height in ems
The icon height was previously set to `@size-150`, which was close
to the message text's line height of `@line-height-medium`.
However, when the size tokens changed from ems to rems, this
relationship no longer holds in smaller font sizes like 14px.

This patch updates the icon height to be `@line-height-medium`
in ems, so it will always match the text height.

Change-Id: Ifdf067adff46caf4b258e5b15202136f39a66d2c