Lookup: always clear the selection when the input changes
Previously, the selection was only cleared if it matched a menu
item, since the Lookup component computes a selectedMenuItem
rather than directly using props.selected in most of its logic.

This means if you have a selection but it doesn't match a menu
item, and the input changes, the selection persists.
This patch updates the logic on input to clear the selection
if props.selected no longer equals the input value.

Bug: T370504
Change-Id: I6f93b10a798175fa4f325aca5bf63d5c050e0414