Generate bi-directional Codex stylesheets using PostCSS RTLCSS
Thus far, we have relied on the RTLCSS library to produce stand-alone
RTL and LTR Codex stylesheets. The assumption was that only one or
the other will ever be loaded at once, and we don't want to ship
unnecessary styles to the client.

This suits the usage in MediaWiki fine, but our system does not
currently support use-cases where a stand-alone Vue app seeks to
allow client-side switching of language and directionality without
forcing a full page reload.

This patch uses PostCSS RTLCSS (a different library from RTLCSS)
to generate a new, experimental bi-directional stylesheet by adding
a new "mode" in our build system. This file contains selectors
scoped to [dir], [dir="ltr"], and [dir="rtl"] which house direction-
specific style rules. Example:

    [dir="ltr"] .cdx-card__thumbnail.cdx-thumbnail {
        margin-right: 12px;
    }

    [dir="rtl"] .cdx-card__thumbnail.cdx-thumbnail {
        margin-left: 12px;
    }

This file should not be used in MW because it will contain redundant
code that is not usable within a given directionality. But it would
provide a way to support both RTL and LTR styles at once in a
single-page application (assuming the `dir` attribute is applied
to a parent element like <html> or <body>).

Bug: T372614
Change-Id: Ie5c44ab6931a6292603a4192ac9d4cc34bd7f100