Lookup: Only clear selection on input if there is one
Previously, the Lookup was clearing its selection on input all
the time, because if the selected value was null it would
not match the new input and thus pass one of the conditionals
that would lead to the modelValue being reset and an
`update:selected` event being emitted. This patch updates that
logic to first check that the selected value is not null.

Bug: T374246
Change-Id: I005d19703b65a15c70dadabd20817fca03ae2dfb