Lookup: Add ARIA `role`` to no result Lookup item
The "no results found" slot of the Lookup component is an `li` tag
inside a `ul`. The `ul` is marked with the ARIA role `listbox`, but
the `li` lacks any role. This is (according to `cypress-axe`) not
ARIA conformant.

Add the `option` role to the `li`, so that the `listbox` has an
`option` child, and that the `li` has a semantic role in the
accessibility tree.

Bug: T376024
Bug: T370057
Change-Id: Iddbd4a83a444b382d0b9318cd1831b42126464c4