Menu: disable Tab keynav selection from multi-select menus
Resolves a bug in multiselect menus.

Before, when you use keyboard navigation to navigate to a menu
item, then hit Enter key to select an item, then Tab to move focus,
the item selected accidentally deselects in the multi-select menus.
This happens because Enter and Tab keys have the ability to select
and deselect menu items.

To resolve this bug in multiselect menus, we've decided to disable
selection via Tab in multi-select menus. Now, when focused on a
multiselect menu, Tab key will only move focus.

In addition:
- Updated the keyboard navigation guidelines to reflect the changes
made to handling Tab key events in menus.
- Update Menu test to ensure Tab selection is disabled in multiselect
menus.

Bug: T374882
Change-Id: I0926b52ca1677d4bcc5d02dfc71f7009b920fafc