ChipInput: On chip click, show chip label or value
Previously, chips could only have a single value, so when a chip
was clicked, that value would become the input value (so it
could be edited). Now that chips can have both a label and a value,
in order to support the multiselect lookup, we need to show the
label if it exists, and the value otherwise.

This means if you click a chip with the label "carrot" and the value
"Q81", the input value will become "carrot", which will then
trigger the lookup to use "carrot" as the search term.

Bug: T375700
Change-Id: I524104d8b76f50615bc7b9336f90d6e04480d6a5