MultiselectLookup: Handle race condition when editing chip
When you click on a chip to edit it, this happens:
1. The chip's label gets moved to the input
2. An input event is emitted, so the parent component fetches
   and passes in menu items
3. The menu opens
4. The menu item related to the chip gets highlighted, because
   the menu component still thinks it's selected
5. The chip's value is removed from the selected array

This is a problem because pressing Enter will not re-select that
item, even though it's highlighted, becuase it was highlighted
via the Menu component and not via the keyboard and highlighted
menu items are only selected on Enter if they were highlighted
via the keyboard.

This patch tells the MultiselectLookup component to wait a tick
before emitting input events to account for this issue.

Bug: T375700
Change-Id: Ic4e5ec6d3dd21fdb58407f60326f3f6f1fbc9c09