ChipInput: Make input background transparent
When ChipInput is used inside MultiselectInput and results take
time to get fetched after user input, the pending animation was
visible outside of the actual `<input>` element in the ChipInput
but not within it. This sets the background color of the input
element to transparent, so it'll always take on the background
color of its parent element.

Bug: T375700
Change-Id: Ibf27a5a3478e6493d373ca387c8311d39ab752fc