Button: trigger actual click on key interaction
The space and enter keys can be used to click a button. We want
the click to go through on keyup instead of keydown, so it won't
matter how long you hold the button down, there will only be a
single click event. However, emitting a Vue event on keyup will
not trigger native browser behavior like submitting a form.

This patch triggers an actual click of the button element, rather
than emitting a click event on keyup.

Bug: T376622
Change-Id: I3dc464b18ccd6fd0621844590142dbe6ac918a2e