InfoChip: CSS-only demo with icon
Add CSS-only InfoChip demo with icon.
- Demonstrate the pre-defined statuses (notice, warning, error,
and success) and how to use a custom icon in a notice chip.
- Refactor InfoChip to ensure that the CSS-only icon is styled
using a Less mixin(`.cdx-mixin-css-icon`). The vue icon class
was updated from `.cdx-info-chip__icon` to
`.cdx-info-chip__icon--vue`. The CSS-only icon class is
`.cdx-info-chip__icon`. This update makes it easier for a user
to add the default status icons in their project as illustrated
in the demos.
(`.cdx-info-chip__icon`) to apply the status styles. This follows
a similar pattern in Message.
- Add Sandbox demo.
- Update the `cdx-info-chip--text` class to `cdx-info-chip__text`
to follow proper BEM conventions.

Bug: T370132
Change-Id: Ifc87683941bbe92b60323480c8cc769dbde7d3dc