docs: Link to the menu group docs
This patch ensures that all menu components (with two exceptions)
have a section on their demo page called "Other features" that
includes a link to the menu with groups demo. Two redundant demos
were also removed from the Combobox page.

The exceptions are Select (because there is a full demo of menu
groups on this page, since optgroups are a common/expected feature
of a select) and TypeaheadSearch (because menu groups are not
enabled for that component).

Bug: T305036
Change-Id: Ibb6b9c89c90694809c35c24241ba4d808d1e3458