docs: Set correct background color
Other than the sidebar, where we had already set
`background-color-base`, the site was using a VitePress color as
the background color, which was especially visible in dark mode
(the header and main content sections were a shade too light).
This patch sets VitePress's main background color token to
`background-color-base` to ensure that the background color is
correct everywhere.

Change-Id: I05fbf90c963eed6d9b594535049b164b4d91df81