Lookup: Ensure input event on selection has correct value
When the selection changes, a watcher runs that sets the input
value to the newly selected item's label (or value, if there is
no label). This will automatically emit an `update:input-value`
event with that new value. We also emit an `input` event with
the same value, in case the optional `inputValue` prop isn't
being used.

We used to emit the `input` event with the value of the input,
an internal ref called `computedInputValue`. After the recent
menu groups change, this ref does not seem to be updated in
time for the input event to be emitted with the proper value.
This patch fixes this by emitting the event with the label
(or value) of the newly selected item instead, since we know
this varible will be up-to-date.

This doesn't resolve the underlying cause, which should be
investigated as part of T378425.

Bug: T378177
Change-Id: Icd4b7ed1e34bf5b2d62f705fdd8c6f4e0cf088d1