Table: Support CSS-only sort icons
We don't actually advertise this on the demo page, but the
MediaWiki TablePager will need to support sort icons. This
patch applies the CSS icon mixin for the 3 sort icons and
adds a sandbox demo of a table with sort icons (the demo
doesn't actually work, but it just needs to prove that the
icons look right).

Bug: T375827
Change-Id: I68ac3c2c1942c2879354b798ad2e659b92a81543