Table: Use colspan and basic table styles for empty state
Codex Table components can display an "empty state" row when
there is no actual data to display. Previously this element was
being styled with Flexbox, but it didn't correctly center within
tables that had more than a single column.

This patch removes the `display:flex` style on this element and
instead adds a `colspan` attribute with a value equal to the number
of columns in the table (zero value here is ok and does not adversely
impact layout).

The empty state content styles have been further modified to use
`height` instead of `min-height` (the former is equivalent to the
latter in display: table), and `text-align: center` and
`vertical-align: middle` instead of the Flexbox equivalents. An
increase in selector specificity was necessary to override the
normal `<td>` element styles used in the component.

Bug: T379004
Change-Id: I3477697f70b206c26890eee3046775898ec82293