InputChip: Disable the delete button in read-only mode
If the parent ChipInput is read-only, deleting chips is not allowed, and
clicking the delete button does nothing. Accordingly, disable the delete
button.

This works around a bug where clicking the delete button on a chip in a
read-only ChipInput submitted the containing form. It's unclear to me
why this happened (or why it only happened in read-only mode), but
disabling the delete button fixes this bug by preventing it from being
clicked at all.

Bug: T382175
Change-Id: Ia80e502dc34e30fedd0ec561a0572bd85ada6f50