docs: Remove remaining demo text
We previously used a special class to style text in the demo pane
that usually showed the current value of a ref. We decided to
remove this text from the demo pane as part of the component
simplification work since it's distracting for non-engineer
users. This patch removes the remaining text and the associated
styles.

Bug: T379299
Change-Id: Ifd6a460ec817196c10e9f051e6cf709cef2d1620