build: Use new @import directive in jsdoc comments
TypeScript 5.5 introduced a new @import directive, which allows us to
simplify this:
    /** @typedef {import('style-dictionary').Platform} Platform */
to this:
    /** @import { Platform } from 'style-dictionary' */

Apply this in all JS files that use jsdoc imports, and merge @import
statements that import from the same package.

As a bonus, use another new TypeScript feature in Menu.vue: allow the
type predicate returned by selectedIsArray() to be inferred
automatically instead of specified manually.

Change-Id: I4bcac4cf21a4f3ba5552b0db9003e25dee15a945