Remove Lucas Werkmeister from .mailmap
<lucas.werkmeister@wikimedia.de> is my work email,
<mail@lucaswerkmeister.de> is my private email. They’re linked to
different Gerrit accounts and should be treated as separate identities.
I’m fine with combining them in AUTHORS.txt (and I think using the
private email there makes sense, as it’s probably slightly more likely
to still be valid in the future), but IMHO they shouldn’t be merged in
the git log.

Change-Id: I7c8dfcbf4a8b80075d6a9a07ac11813eabf16d49