build: Relax the peerDependency on vue
We were requiring that Codex was used with the exact version we use:
only Vue 3.5.13 is accepted, 3.5.14 or 3.6.x would not be permitted.
This is unworkable because it makes updating the Vue version in
MediaWiki (and other packages using Codex) difficult.

Instead, change the peerDependency to ^3.5.13, which permits 3.5.13 or
any later version without breaking changes (i.e. anything between 3.5.13
and 4.0.0). This makes it easier for downstream packages to update their
Vue version independently, or update it in a separate commit before
updating to a new Codex version that requires a later version of Vue.

Keep the version that Codex itself uses internally pinned to 3.5.13.

Change-Id: I2b9e5dba968a1e90f8dea25d803aa8cb22d5aad1