icons, docs: Remove `xmlns:xlink` namespace where unnecessary
While debugging something else (see T361670#10452424) I was
wondering why all icons include that extra namespace. Only 4 icons
(from currently 284) actually use it.

I believe this might have a surprisingly big impact. While the string
is not long (only 43 extra characters) and it's the same all the time
(which means it should compress very well) it's probably not needed
at all in the vast majority of situations (when none of the 4 icons
is involved).

Also updating snapshots and documentation with included and now reduced SVGs.

Bug: T358062
Change-Id: I6e9102731053351ba85f183099851c4455b87ef9