demos: Make `tbody` explicit in `table`s
While not strictly necessary, it is a common practice
to omit the `tbody` tag in `table`s.
This is because the `tbody` tag is implicit in the browser and
is automatically added by the browser when it is not present.
This change makes the `tbody` tag explicit in all `table`s
in the component demos to circumvent a VitePress warning about
potential hydration errors.

Change-Id: Iae693d765c7ee8b41314d26baf72d711fc5267e0