styles, Button: fix the progressive primary button's icon
Update the CSS-only icon's background color in progressive
primary buttons so that it appears in both light mode and
dark mode by setting the color to `@color-inverted-fixed`.

This fixes a bug introduced in Codex v1.18.0 where we made
updates to button styles. The bug affected CSS-only buttons,
link buttons, and checkbox hack buttons which are displayed on
our Sandbox page. The icon was "missing" because the icon color
rule was removed in a previous patch which made the icon have
the same background color as the button.

In addition, organize the Button styles to ensure psuedo-classes
are listed before children elements such as icon.

Bug: T382758
Change-Id: I9cf8f6a3e5be891921fe53852e3e1146a0248bd1