Menu: Re-work how IDs are generated
Currently, the Menu component calls Vue's `useId` composable
inside the `computedMenuEntries` computed prop. This is not safe;
Vue.js composables are only intended to be called during setup,
not at runtime in the body of computed properties or watchers.
Broken behavior may result when trying to interact with Menu
items that have been added after the component's setup function
runs (i.e. in Lookup components, where the Menu is populated
by the results of a fetch request); users of assistive technology
may be impacted in these situations. When Vue is running in dev
mode, console warnings can be seen in these scenarios.

This patch introduces a new approach to ID generation for MenuItems
inside of Menu components that avoids using useId outside of setup,
while also ensuring that MenuItems added during runtime receive
unique IDs.

First, each Menu component calls useId once, at the beginning of
setup, to generate a unique ID prefix.

Next, a reactive counter is set up that is incremented each time
a new MenuItem ID gets generated.

A Map object is also initialized here, to store and retreive
the ids which have been generated thus far. IDs are keyed to
the `value` of MenuItems, or to a string derived from the label
of Group items (since they have no value).

If a new Menu item is added, it gets a unique ID that follows
the format of `${ menuInstanceID }-{ idCounter.value }`, ex:
"v-0-6" (for the 7th item of the first Menu in a given app).
This is handled whenever the body of the computedMenuEntries
computed prop is executed.

Finally, a watcher pays attention to changes in the menuItems
prop and ensures that stale entries are deleted from the Map.

Bug: T385099
Change-Id: Ic8e68bfcd93eb947edac0d50a091761d46eedc19