Menu: Add `tabindex="-1"` to `ul` element
The `<ul>` inside the Menu component is never meant to be directly
focused – instead, the user focuses on the triggering element,
then uses arrow keys to navigate through the menu items.

However, when using the tab key to move focus from the triggering
element to the next interactive element on the page, the menu
may become momentarily focused until it is hidden (since the menu
is meant to close on blur). Focus is then moved to the body
element. In some browsers, pressing tab again will focus on the
appropriate interactive item after the triggering element, but
this is not implemented in all browsers (e.g. Firefox).

To avoid this issue, we explicitly remove the `<ul>` from the tab
order by setting `tabindex="-1"` so it cannot be focused.

Bug: T383867
Change-Id: I93f0ce34f87ebef4ea328e3e5e1655ebce55b5a1