docs: Don't break everything when there's a generic type
One of the Popover props uses generic types, which contain angle
brackets, which breaks the markdown parser that builds the
props tables for component pages. This patch updates our code
that processes types (usually to add links to the Types and
Constants page) to slap some backticks around generic types
and bail early. This is a stop-gap solution and we should try
to provide better support for generic types in the future.

Change-Id: I7bc9ca5e51ce5fd398a896fff29688b3cc3833d6