Tooltip: Improve ID generation
The Tooltip component is actually implemented as a directive, so it
cannot make use of Vue's new useId() composable that other Codex
components rely on for this sort of thing.

This change addresses the above problem in two parts. First, it
introduces a general-purpose hashing function ("generateHashId")
that is based on the String.hashCode function from Java, which
has been widely adapted. The hashing function is extensively
documented because it relies on some concepts which may be
unfamiliar to many Codex users and developers. This hash is
appropriate for usage in situations like ID generation, caching,
etc., but it is not cryptographically safe.

The Tooltip component uses this hashing function along with a new
Map object that keeps track of every Tooltip instance that is
generated (Tooltips with the same content are reflected by
incrementing a counter in a map entry's value). Together, this
sytem is used to generate reasonably unique, deterministic,
short IDs in a performant way, that is SSR-safe (IDs generated
on the server and client will match).

Change-Id: I2903a012465a32984eafc9b8b7d8246c5c8eff7e