Popover: Make `anchor` loosely required
The `anchor` prop is required so we can position the Popover relative
to it. However, since a template ref gets passed in, it will be
undefined when the component is created but not yet mounted. This
causes a Vue warning to be logged since the prop is required yet
undefined.

This patch:
- Makes the `anchor` prop not required and defaulting to `null`
- Adds prop documentation describing the prop's use and that it
  is, effectively, required
- Adds a console warning if the prop is still `null` on mount

Bug: T386485
Change-Id: I42fcd5a591e6d55845ae392f534f23d8cc7892fd