Popover: fix console warning
Ensure the console warning to provide an anchor prop is
displayed when the component mounts and then on next tick
check if the anchor prop was provided.
When the component mounts, the anchor is null and the
console warning displayed in the browser whether the anchor
was provided or not. Adding `nextTick()` ensures that the
code waits for the next DOM update then checks if anchor
was provided.

Change-Id: I3be88744bda22cc919a6a343f6c7348712af7d1e