Popover: Change "positionConfig" prop to "placement"
Instead of passing rich "config" objects with multiple
distinct keys as props to our components, we should destructure
things where possible and pass in props one at a time.

Currently the only aspect of FloatingUI's PositionConfig
interface we need for this component is the Placement value.
This change replaces the "placementConfig" Popover prop with
a "placement" prop.

This brings Popover more in line with Tooltip, which also accepts
a placement option.

Also updates the Popover demo to support configuration of the
placement option in the interactive demo.

Change-Id: I9f29ef458cbff2f30acf268facf8a918cc427ca7