Popover: Set max-width instead of width
Popovers should be as wide as their content necessitates, while
staying within the min- and max-widths defined in the design.

This patch changes FloatingUI's size middleware to set `max-width`
instead of `width`. The max-width is the available space (up to
the limit of 512px), or the min-width of 256px if that's larger.

Bug: T388709
Change-Id: Iab10afd7bbcf9652687ab6b8b2edaa4d9b48b4e2