build: Deduplicate code for building mode outputs
The code for building the dark mode, small mode, large mode and x-large
mode outputs is identical except for the name of the mode, which
appears in half a dozen places. Rewrite this as a loop over the mode
names.

If we want to move the font-related modes to their own directory, or add
a prefix to them (e.g. .cdx-mode-font-small()), then this code would
need to be tweaked a bit to add that information to the modes array that
the loop loops over.

Change-Id: I43f719b3035dca7ae9b88d7272c9dd174ceadfa5