demos: Add typography demo sandbox page
This aims to enable us to test the coming font mode feature in Codex.
This is pretty rudimentary for the time being, but getting it in will
enable us to start testing the change by the font mode in a
self-contained environment, with relying only on HTML and CSS with
Codex design tokens and no framework bleed-in from MediaWiki, skins,
extensions or VitePress.

Bug: T386298
Change-Id: Id4890b55a6e74e2bedf09bc592fd6d1e8e98d3f1