styles, docs: Harden design token usage in calculations
As design token values could be numeric values with
units, but also upcoming CSS custom properties,
it's useful to rely on CSS `calc()` function to
ensure that the values are correctly calculated.
This will also bring us closer to removing Less parser
in future and rely on CSS standard only.

Also providing an appendix to the font mode ADR in
order to clarify the Less parser behaviour better.

Bug: T389234
Change-Id: I480d77a25f55c35fd7d15605da785b3c4d72b02f