Accordion: add support for optional v-model prop
This change updates the Accordion component so that it can
optionally accept a v-model binding. This allows the
component's "open" state to be controlled programmatically
from a parent scope.

If the optional v-model binding is not provided, then the
component behaves as it did before; the native `<details>`
element handles the open state in response to toggle events,
independently of the larger Vue application. As before, an
"open" attribute can be provided to control the initial state
of the `<details>` element.

If *both* an "open" attribute *and* a v-model binding are
provided, then the "open" attribute is used to provide the
component's initial state.

Two new events have also been added to the Accordion
component. If the optional v-model binding is used, then
"update:modelValue" events are emitted normally whenever the
bound value changes. This event is *not* emitted if no binding
has been provided.

In addition to "update:modelValue", the Accordion also emits
a "toggle" event that fires whenever the internal `<details>`
element toggles; this event is emitted regardless of the
v-model binding.

Both new events provide the boolean state of the disclosure
element as a payload.

Tests and demos have been updated accordingly.

Bug: T391816
Change-Id: Ia01b74807f24a496975739207a5727b48e191753