[BREAKING CHANGE] tokens: Add new Less mixin files and remove legacy file
This patch re-organizes some of the build artifacts in the
tokens file for greater consistency and clarity, now that
we are shipping multiple modes.

New files:

* theme-wikimedia-ui-mixin-light.less
* theme-wikimedia-ui-mixin-medium.less

These files make the default values for color and font modes
available as Less mixins that can be called anywhere:
.cdx-mode-light() and .cdx-mode-medium(), respectively.

Use these mixins to reset a single mode to the default
value without changing other modes; they will also be useful
if a skin like Monobook uses something like the small font
mode as a default (in which case the "medium" option can be
invoked as needed just like all other options).

Changed files names:

* theme-wikimedai-ui-reset.less -> theme-wikimedia-ui-mixin-mode-reset.less

Deleted files:

* theme-wikimedia-ui-legacy.less

Finally, the README file for codex-design-tokens has been updated
to include a list of all the various build artifacts along with
guidelines for which files to use in what circumstances.

Bug: T392030
Change-Id: Iec1ff023cc56aeb8c582fb82b3a7ba4e5c1a68a0