Accordion: Use flexbox instead of positioning for action button
Previously the Accordion component used `position: absolute` to
ensure that the optional "action button" was always lined up on
the right-hand side of the Accordion header. This approach also
necessitated setting a `position: relative` rule on the entire
Accordion component.

The recent changes around font modes have shown this solution
to not be very flexible, and it's easy to wind up in situations
where the button is either stretched, mis-aligned, or collides
with long text content.

This patch introduces a flexbox-based solution to align the
action button in the desired location without changing any of
the component's markup. Some redundant CSS rules have been removed.

Bug: T392932
Change-Id: I0d080d11abeaccb2ff69201c88d56383990c91ff