docs: Sync VitePress's `isDark` ref with color mode switcher
We recently removed VitePress's built-in color mode switcher in
favor of our own custom "Appearance" menu and are now storing
the current color mode in localStorage. This patch ensures that
VitePress's `isDark` ref is kept in sync with the selected color
mode. This ensures `isDark` can be used throughout the site and
is reactive to color mode changes.

This fixes a bug on the "Colors" design tokens page where color
swatches and values were no longer updating when the color mode
changed.

Bug: T393389
Change-Id: I9f2347f8acc49ff6fec395bcc587174249b7c782