tokens: introduce `line-height-content` token
This change introduces a new application token, `line-height-content`.

Unlike other `line-height` tokens (which have specific values measured
in `rem` units), the `line-height-content` token's value is a unitless
ratio. This token is intended to be used in places where user-generated
content (with arbitrary `font-size` styling) may appear; in this case,
a unit-less value provides maximum flexibility. The exact ratio used
will still vary based on the current font mode.

In order to bypass the existing `shouldUseRelativeSize` matcher (which
needs to still pick up all other "line-height" tokens), a new property
called `unitless` has been added to the token definition in application.json.

In the future, if any other tokens are introduced which need to be handled
as an exception to a typical relative/absolute unit transformation, this
convention could be followed again; just set `unitless: true` in the
token definition, and ensure that the relevant matcher checks for this
property.

Bug: T394722
Change-Id: I331bc5b5e96124bfcf473e47ab95ec2eb8df8808