Menu: Don't close when scrollbar is clicked
We encourage parent components that use Menu to close the menu when the
input they're attached to is blurred. However, we don't want a click in
the menu itself to cause the menu to close. To achieve that, we prevent
mousedown in MenuItem, so that clicking menu items doesn't steal the
focus from the input and close the menu.

But when the menu has a scroll bar, clicking that scroll bar will cause
a mousedown event on the scrollable `<ul>` in Menu, not on any of the
MenuItems. To prevent this from closing the menu, also prevent mousedown
on the `<ul>` in Menu.

Bug: T395936
Change-Id: Ia281020c5ac0f944cbc9cde0dc4089227db4788e