Select: Add hidden HTML form element when a name is provided
Add a input element with type "hidden" to the Select component. It only
exists when a name is provided. It's value is set to the currently
selected option's value. To test this component two tests were added:
a test to check the hidden input reports it's value correctly to the
form and a test to check the reported value updated with a change in the
selection. All snapshots were updated to reflect the new hidden input.

Bug: T365782
Change-Id: Id4c3b765d8c379cd5495625373dba255f0e07152