Image, styles: unify placeholder, loading, and error states
This patch updates the template logic that conditionally renders
the loading, error, and placeholder states. Also, removes the
custom error and loading elements from the template. When an image
fails to load, the alternative text is displayed. When loading,
the placeholder is diplayed for now until the Skeleton loading
element is implemented.

Template logic updates:
- Display the placeholder during the initial loading state or
when no image src is provided.
- Display the image element for success and error cases (broken
image with alternative text).

Style updates:
- Restructure Less code and nest styles.
- Change class names to follow Codex writing conventions (eg. BEM).
- Apply subtle border and border radius to placeholder state.
- Remove custom loading and error div element styles.
- Remove border styles when in an error state.
- Hide the image element with `visibility: hidden` when image is
loading. Both elements (image and placeholder) are rendered  when
loading. The image needs to be in the DOM during loading, to trigger
the `load` event. The `load` event updates `isLoaded` state to true.
- Absolutely position the placeholder when the placeholder is
preceded by the image element eg. during loading state.
- Dynamically apply aspect ratio classes to the placeholder
(similar to how it's applied to the image). This ensures the aspect
ratio styles are consistent between the image and placeholder.
- Center the Image demos on the component demo page.
- When there is an error, the alt text is 12px (similar to caption
font size).

Other:
- Update test files
- Remove Image error example from the component page. Note: there's
an example on the Sandbox page.

Bug: T381525
Change-Id: I25f4fc83b78bd2efd738f0cf8ab491636094d502